feat(git): remove Git `core.editor` config
Let Git use system's default editor. On Unix, use the `VISUAL`
environment variable. On Windows, it'll use `core.editor` config in
`C:/Program Files/Git/etc/gitconfig`. This is so that it works on both
Unix and Windows. If this is set to `vim`, on Windows it'll use the one
bundled with Git instead of the user's installed vim, which doesn't have
any plugins so the user's `vimrc` will break vim.