fix(vim): ensure powershell exists before running during theme detection
When running in a Docker container using WSL 2, it still detects that we
are in WSL 2 and will attempt to detect the Windows host's color scheme
using `powershell.exe`. The command will fail because powershell doesn't
exists rather than not finding the `AppUseLightTheme` registry, thus
falsely detects that we are running dark mode. Hence, we must check if
`powershell.exe` exists before running it.