fix(vim): don't organise import on save for Windows
We can't run organize on save on Windows, thus the mapping `<space>i` is
added for organizing imports.

This is because on Windows, it seems that the file finishes writing
before the organize imports operation completes. As a result, the saved
file does not include the final organized import changes, and the editor
still considers the file modified, requiring it to be saved again,
indefinitely. Hence, we can't run organize on save on Windows and the
mapping `<space>i` is added for organizing imports.